Split pure formatting out of the capture orchestrator, and move a rule set to MoonBit
Step 1 — extract pure from IO. `component-from-image.ts` was 1,867 lines
holding two unrelated things: a 729-line Playwright orchestrator and 704
lines of pure Markdown formatting over the report it produced. The block
was contiguous, with zero async calls and zero IO, so it moves out whole
as `component-report-format.ts`.

The point is not tidiness. Five tests for `renderReportMarkdown` already
existed and were pure, but they imported the module that statically
imports Playwright: 562ms of module loading against 85ms for the formatter
alone. Moving them alongside the code they test takes that file from 777ms
to 254ms and out of the browser dependency entirely. Same eleven tests,
now split by what they actually need.

Step 2 — move a rule set. `deriveInteractionIssues` is eleven
keyboard-a11y rules over an array of records with an optional nested
record, which is exactly what the positional boundary could not express;
those rules were in TypeScript because there was no way to pass their
input, not because that was where they belonged. They are now
`markup-core/interaction_issues.mbt`, called through the JSON boundary,
with all twenty existing tests passing unchanged.

Three things that keep the handler honest, now documented as the pattern:

- Ids out, not prose. MoonBit returns the issue id, severity, element
  position and the values the wording interpolates; the sentences are a
  TypeScript table. Rules and wording change for different reasons.
- Send what the rules read. `ariaDelta` is a Record of tuples and the
  rules only ask whether it is empty and whether it mentions `expanded`,
  so two booleans cross the boundary rather than the map — otherwise every
  future change to that type becomes a change in MoonBit.
- Identify elements by position, not by a field they carry. Reporting the
  element's own `index` looked equivalent and was not: it is a discovery
  index a caller may repeat, and the test fixtures do, so every finding
  came back labelled with the last element sharing the value. The existing
  tests caught it, which is why moving a rule set wants its own safety net
  before the move rather than after.
